+++ ile_de_france/src/main/java/org/eqasim/ile_de_france/IDFConfigurator.java
@@ -1,23 +1,6 @@
 package org.eqasim.ile_de_france;
 
 import org.eqasim.core.simulation.EqasimConfigurator;
-import org.matsim.api.core.v01.Id;
-import org.matsim.api.core.v01.Scenario;
-import org.matsim.api.core.v01.population.Person;
-import org.matsim.households.Household;
 
 public class IDFConfigurator extends EqasimConfigurator {
-	static public void adjustScenario(Scenario scenario) {
-		EqasimConfigurator.adjustScenario(scenario);
-
-		for (Household household : scenario.getHouseholds().getHouseholds().values()) {
-			for (Id<Person> memberId : household.getMemberIds()) {
-				Person person = scenario.getPopulation().getPersons().get(memberId);
-
-				if (person != null) {
-					copyAttribute(household, person, "carAvailability");
-				}
-			}
-		}
-	}
 }